+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [2.11.7]
 - Fix diagonal movement speed for enemies
+- Improvements in `moveOnlyMapArea`
 
 # [2.11.6]
 - Update Flame to `1.5.0`